add wifi adapter

+++ src/model/co2_data.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct CO2Data {
-    uint16_t co2;
-    uint16_t tvoc;
+    float co2;
+    float tvoc;
 };
