+++ src/model/tp_data.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct TPData {
-    uint16_t temp;
-    uint16_t pressure;
+    float temp;
+    float pressure;
 };
